Server~/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116

aiohttp==3.11.11
aioice==0.9.0
aiortc==1.6.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
distro==1.9.0
dnspython==2.6.1
frozenlist==1.5.0
numpy==1.26.4
opencv-python==4.8.1.78
pillow==10.4.0
pydantic==1.10.9
pydantic_core==2.27.2
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-extensions
urllib3==2.3.0
yarl==1.18.3
soundfile==0.13.1
SpeechRecognition==3.14.1

torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0

openai==1.61.1
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e

accelerate==0.21.0
sentencepiece==0.1.99
einops==0.6.1


soundfile==0.13.1
ultralytics==8.3.105
